Add files via upload
In this notebook, I want to show two types of dimensionality reduction for tabular data: PCA and Autoencoders.

I use the wine dataset to show how Variational Autoencoder (VAE) with PyTorch on tabular data works and compare it to the classic PCA approach. I use the PCA/VAE to reduce the dimensionality of the dataset, in this case, don to 3 Variables (embeddings). I then plot the embeddings in a 3D graph to show how VAE is similar to a PCA but works in a non-linear way.